feat: organize project structure by creating index files for schemas, controllers, services, routes, middleware, and validators
