+++ src/config/dotenv.ts
@@ -1,0 +1,4 @@
+import dotenv from 'dotenv';
+
+// Load environment variables from .env file
+dotenv.config();
