+++ src/routes/index.ts
@@ -1,0 +1,2 @@
+export * from './authRoutes';
+export * from './messageRoutes';
