--- src/types/server/http.ts
@@ -1,8 +1,0 @@
-/**
- * Represents an error response.
- */
-export interface ErrorResponse {
-  status: 'error';
-  message: string;
-  code?: string;
-}
